Created script to insert tables in PostgresQL
Began creating the tables for PostgresQL. All datatypes and surface level relationships (keys and references) should be being referenced right now.

REGARDING NEW TABLE Overarching_Items: We have deviated from our database design document and added an extra table Overarching_Items. This tbale is responsible for storing the items that contain other items, such as the bowl, the plate, the large plate, the family meal, and the panda cub meal. This keeps track of the original pricing of those items, as well as the number of different items that it keeps track of (ex. bowl will have 1 side and 1 entree). This solves the issue of keeping track of the overarching menu item separate from the menu items (which keeps track of the individual entrees, sides, appetizers, etc instead).